processing/requirements.txt
annotated-types==0.6.0
click==8.1.8
grpcio==1.71.0
grpcio-tools==1.71.0
iniconfig==1.1.1
packaging==24.2
pillow==11.3.0
pluggy==1.5.0
protobuf==5.29.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.1.0
redis==3.5.3
setuptools==78.1.1
typing-inspection==0.4.0
typing_extensions==4.12.2
wheel==0.45.1
